Renaming some vision test fixture imports.
Done mostly via:

$ git grep -l 'unit_tests.vision._fixtures' |
> xargs sed -i s/'unit_tests.vision._fixtures'/'unit_tests._fixtures'/g

Then fixed up some indents after the fact.
